functions-datapipeline/function-tc-apicall/requirements.txt
functions-framework==3.*
requests==2.31.*
google-cloud-storage==2.10.*
pandas==2.1.*